Refactoring Main process logic
